Move Documentation styles to clearer homes
None of these were really most directly associated with the /docs pages, which is how we use the name "Documentation"
