+++ src/components/Navigation/SharedStyles.js
@@ -11,6 +11,7 @@ import {
 
 import Logo from "assets/icons/stellar-logo.svg";
 
+import { Column } from "basics/Grid";
 import { Image } from "basics/Images";
 import { H2 as BasicH2 } from "basics/Text";
 
@@ -80,6 +81,11 @@ const LogoEl = styled(Logo).attrs({ width: 100, height: 24 })`
 `};
 `;
 
+export const NavColumn = styled(Column)`
+  position: sticky;
+  height: 100vh;
+  top: 0;
+`;
 export const NavDivider = styled.span`
   display: inline-block;
   width: 0.0625rem;
